// src/redux/types/repository.tsx
export interface IRepositoryOwner {

  /**
   |--------------------------------------------------------------------------
   | Id. (Required)
   |--------------------------------------------------------------------------
   */
  id: number;

  /**
   |--------------------------------------------------------------------------
   | Login. (Required)
   |--------------------------------------------------------------------------
   */
  login: string;

  /**
   |--------------------------------------------------------------------------
   | Avatar url. (Not required)
   |--------------------------------------------------------------------------
   */
  avatar_url?: string;

}

export interface IRepository {

  /**
   |--------------------------------------------------------------------------
   | Id. (Required)
   |--------------------------------------------------------------------------
   */
  id: number;

  /**
   |--------------------------------------------------------------------------
   | Name. (Required)
   |--------------------------------------------------------------------------
   */
  name: string;

  /**
   |--------------------------------------------------------------------------
   | Full name. (Required)
   |--------------------------------------------------------------------------
   */
  full_name: string;

  /**
   |--------------------------------------------------------------------------
   | Description. (Not required)
   |--------------------------------------------------------------------------
   */
  description?: string;

  /**
   |--------------------------------------------------------------------------
   | Url. (Required)
   |--------------------------------------------------------------------------
   */
  html_url: string;

  /**
   |--------------------------------------------------------------------------
   | Owner. (Required)
   |--------------------------------------------------------------------------
   */
  owner: IRepositoryOwner;

}